+++ RarelySimple.AvatarScriptLink/Objects/OptionObject2015.cs
@@ -112,7 +112,7 @@ namespace RarelySimple.AvatarScriptLink.Objects
         /// <param name="errorCode"></param>
         /// <param name="errorMessage"></param>
         /// <returns></returns>
-        public new OptionObject2015 ToReturnOptionObject(int errorCode, string errorMessage) => (OptionObject2015)OptionObjectHelpers.GetReturnOptionObject((IOptionObject2015)this, errorCode, errorMessage);
+        public new OptionObject2015 ToReturnOptionObject(double errorCode, string errorMessage) => (OptionObject2015)OptionObjectHelpers.GetReturnOptionObject((IOptionObject2015)this, errorCode, errorMessage);
 
         /// <summary>
         /// Returns a <see cref="string"/> with all of the contents of the <see cref="OptionObject2015"/> formatted as XML.
